Visualize: Bundle digraph code using webpack
