+++ M2/Macaulay2/packages/Visualize/digraph-example.html
@@ -16,7 +16,7 @@
       menuOpen = true;
       dataData = [[0, 1, 1, 0, 0, 0, 0], [0, 0, 0, 1, 1, 0, 0], [0, 0, 0, 0, 1, 1, 0], [0, 0, 0, 0, 0, 0, 1], [0, 0, 0, 0, 0, 0, 1], [0, 0, 0, 0, 0, 0, 1], [0, 0, 0, 0, 0, 0, 0]];
       labelData = ["1", "2", "3", "4", "5", "6", "7"];
-      portData = 8081; // initializes port from user
+      portData = 8080; // initializes port from user
       forceOn = true;
       labelsOn = true;
       forceCharge = -1500;
@@ -59,167 +59,7 @@
 
   <script src="js/visCommon.js"></script>
   <script src="js/BootSideMenu.js"></script>
-  <script src="js/d3.v3.min.js"></script>
   <script src="js/visDigraph2d.js"></script>
-
-  <script>
-      
-    console.log(dataData);
-    console.log(labelData);
-    console.log(8081);
-      
-    // Initialize clipboard.js.
-    var clipboard = null;
-      
-//    $('#side').BootSideMenu({side:"right"});
-    $('#side').BootSideMenu({side:"right", closeOnClick: false, width: "230px"});
-
-    // When the side menu bar is opened or closed (i.e, when the "toggler" div is clicked), resize the svg appropriately so that nodes do not go behind the side menu.
-    document.getElementsByClassName("toggler")[0].addEventListener("mousedown", function() {
-            menuOpen = !menuOpen;
-            updateWindowSize2d();    
-    }, false);  
-    
-    //document.getElementById("canvasElement2d").style.width = window.innerWidth;
-    //document.getElementById("canvasElement2d").style.height = window.innerHeight;
-      
-    window.addEventListener("resize", updateWindowSize2d, false);
-
-    // Initialize sliders
-    var chargeSlider = document.getElementById('charge-slider');
-    noUiSlider.create(chargeSlider, {
-        start: [1500],
-        //tooltips: true,
-        range: {
-			'min': [0],
-			'max': [6000]
-		}
-    });
-    //chargeSlider.noUiSlider.on('update', updateForceCharge);
-      
-    var linkDistSlider = document.getElementById('linkdist-slider');
-    noUiSlider.create(linkDistSlider, {
-        start: [100],
-        //tooltips: true,
-        range: {
-			'min': [0],
-			'max': [400]
-		}
-    });
-    //chargeSlider.noUiSlider.on('update', updateForceLinkDist);
-      
-      
-    $(document).ready(function(){
-
-      $("#editToggle").on("click", function(){
-        if(curEdit) {
-          $(this).html("Enable editing");
-          curEdit = !curEdit;
-          disableEditing();
-        } else {
-          $(this).html("Disable editing");
-          curEdit = !curEdit;
-          enableEditing();
-        }
-      });
-        
-      $("#labelToggle").on("click", function(){
-        if(!labelsOn) {
-          $(this).html("Hide labels");
-          labelsOn = !labelsOn;
-          showLabels();
-        } else {
-          $(this).html("Show labels");
-          labelsOn = !labelsOn;
-          hideLabels();
-        }
-      });
-    
-      $("#highlightToggle").on("click", function(){
-        if(curHighlight) {
-          $(this).html("Highlight neighbors");
-          unHighlightAll();
-          curHighlight = !curHighlight;
-        } else {
-          $(this).html("Disable highlighting");
-          enableHighlight();
-          curHighlight = !curHighlight;
-        }
-      });
-
-      $("#reset").on("click", resetGraph);
-      
-      $("#forceToggle").on("click", toggleForce);
-      
-      $("#exportTikz").on("click", function() {
-        exportTikz();
-      });
-
-      // Begin browser-M2 communication.
-      // For each item, a line in the function `onclickResults()`
-      // located in `visGraph2d.js` must be added. 
-      // 
-      // If you wish to delete the text to make it vanish when the graph is edited
-      // search for 'menuDefaults()'.
-
-      // Checks to see if user's graph has an Eulerian trail
-      $("#hasEulerianTrail").on("click", function() {
-        clickTest = "hasEulerianTrail";
-        makeCorsRequest('POST','http://localhost:'+portData+'/hasEulerianTrail/', digraph2M2Constructor(nodes,links));
-      });
-        
-      // Checks to see if users graph is Eulerian
-      $("#isEulerian").on("click", function() {
-        clickTest = "isEulerian";
-        makeCorsRequest('POST','http://localhost:'+portData+'/isEulerian/', digraph2M2Constructor(nodes,links));
-      });      
-
-      // Checks to see if users graph is strongly connected
-      $("#isStronglyConnected").on("click", function() {
-        clickTest = "isStronglyConnected";
-        makeCorsRequest('POST','http://localhost:'+portData+'/isStronglyConnected/', digraph2M2Constructor(nodes,links));
-      });      
-
-      // Checks to see if users graph is weakly connected
-      $("#isWeaklyConnected").on("click", function() {
-        clickTest = "isWeaklyConnected";
-        makeCorsRequest('POST','http://localhost:'+portData+'/isWeaklyConnected/', digraph2M2Constructor(nodes,links));
-      });      
-        
-      // Ends the browser session and outputs the information to M2
-      $("#endSession").on("click", function() {
-        if(activeSession) {
-          //$(this).html("Session Terminated");
-          //$(this).html(<p style="color: #ffffff; background-color: #ff0000">Session terminated</p>);
-          document.getElementById("endSession").style.color = 'white';
-          document.getElementById("endSession").style.backgroundColor = 'red';
-          document.getElementById("endSession").innerHTML = "Session terminated";
-          makeCorsRequest('POST','http://localhost:'+portData+'/end/',digraph2M2Constructor(nodes,links));
-          activeSession = !activeSession;
-        } else {
-          return;
-        }
-      });            
-
-      initializeBuilder();
-      disableEditing();
-    });
-
-
-    function toggleForce() {
-      if (forceOn) {
-        setAllNodesFixed();
-        document.getElementById("forceToggle").innerHTML = "Turn on force";
-      }
-      else {
-        setAllNodesUnfixed();
-        document.getElementById("forceToggle").innerHTML = "Turn off force";
-      }
-      forceOn = !forceOn;
-    }
-    
-    </script>
-
   </body>
